Tear down ObserveRIB streaming RPC when RIB is destroyed
