+++ routingtable/locRIB/loc_rib.go
@@ -344,3 +344,10 @@ func (a *LocRIB) ReplaceFilterChain(filter.Chain) {
 func (a *LocRIB) RefreshRoute(*net.Prefix, []*route.Path) {
 
 }
+
+// Destroy tells all clients that this LocRIB is not to be used anymore (this can happen when RIS loses a BMP connection)
+func (a *LocRIB) Destroy() {
+	for _, c := range a.clientManager.Clients() {
+		c.Destroy()
+	}
+}
